+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * `records` mode for tabular files
-* `.pickle`
+* `.pickle`format
+* optional pandas import
 
 ## [0.0.1] - 2022-06-27